Optimization Solutions

SciMLBase.OptimizationSolution — Type

struct OptimizationSolution{T, N, uType, C<:SciMLBase.AbstractOptimizationCache, A, OV, O, S, ST} <: SciMLBase.AbstractOptimizationSolution{T, N}

Representation of the solution to a non-linear optimization defined by an OptimizationProblem

Fields

u: the representation of the optimization's solution.

cache::AbstractOptimizationCache: the optimization cache` that was solved.

alg: the algorithm type used by the solver.

objective: Objective value of the solution

retcode: the return code from the solver. Used to determine whether the solver solved successfully or whether it exited due to an error. For more details, see the return code documentation.

original: if the solver is wrapped from an alternative solver ecosystem, such as Optim.jl, then this is the original return from said solver library.

solve_time: Solve time from the solver in Seconds

stats: statistics of the solver, such as the number of function evaluations required.