OptimizationFunction

SciMLBase.OptimizationFunction — Type

struct OptimizationFunction{iip, AD, F, G, H, HV, C, CJ, CH, HP, CJP, CHP, O, EX, CEX, SYS, LH, LHP, HCV, CJCV, CHCV, LHCV} <: SciMLBase.AbstractOptimizationFunction{iip}

A representation of an objective function f, defined by:

\[\min_{u} f(u,p)\]

and all of its related functions, such as the gradient of f, its Hessian, and more. For all cases, u is the state and p are the parameters.

Constructor

OptimizationFunction{iip}(f, adtype::AbstractADType = NoAD();
                          grad = nothing, hess = nothing, hv = nothing,
                          cons = nothing, cons_j = nothing, cons_h = nothing,
                          hess_prototype = nothing,
                          cons_jac_prototype = nothing,
                          cons_hess_prototype = nothing,
                          syms = nothing,
                          paramsyms = nothing,
                          observed = __has_observed(f) ? f.observed : DEFAULT_OBSERVED_NO_TIME,
                          lag_h = nothing,
                          hess_colorvec = __has_colorvec(f) ? f.colorvec : nothing,
                          cons_jac_colorvec = __has_colorvec(f) ? f.colorvec : nothing,
                          cons_hess_colorvec = __has_colorvec(f) ? f.colorvec : nothing,
                          lag_hess_colorvec = nothing,
                          sys = __has_sys(f) ? f.sys : nothing)

Positional Arguments

f(u,p,args...): the function to optimize. u are the optimization variables and p are parameters used in definition of

the objective, even if no such parameters are used in the objective it should be an argument in the function. This can also take any additional arguments that are relevant to the objective function, for example minibatches used in machine learning, take a look at the minibatching tutorial here. This should return a scalar, the loss value, as the first return output and if any additional outputs are returned, they will be passed to the callback function described in Callback Functions.

adtype: see the Defining Optimization Functions via AD section below.

Keyword Arguments

grad(G,u,p) or G=grad(u,p): the gradient of f with respect to u. If f takes additional arguments then grad(G,u,p,args...) or G=grad(u,p,args...) should be used.

hess(H,u,p) or H=hess(u,p): the Hessian of f with respect to u. If f takes additional arguments then hess(H,u,p,args...) or H=hess(u,p,args...) should be used.

hv(Hv,u,v,p) or Hv=hv(u,v,p): the Hessian-vector product $(d^2 f / du^2) v$. If f takes additional arguments then hv(Hv,u,v,p,args...) or Hv=hv(u,v,p, args...) should be used.

cons(res,x,p) or res=cons(x,p) : the constraints function, should mutate the passed res array with value of the ith constraint, evaluated at the current values of variables inside the optimization routine. This takes just the function evaluations and the equality or inequality assertion is applied by the solver based on the constraint bounds passed as lcons and ucons to OptimizationProblem, in case of equality constraints lcons and ucons should be passed equal values.

cons_j(J,x,p) or J=cons_j(x,p): the Jacobian of the constraints.

cons_h(H,x,p) or H=cons_h(x,p): the Hessian of the constraints, provided as an array of Hessians with res[i] being the Hessian with respect to the ith output on cons.

hess_prototype: a prototype matrix matching the type that matches the Hessian. For example, if the Hessian is tridiagonal, then an appropriately sized Hessian matrix can be used as the prototype and optimization solvers will specialize on this structure where possible. Non-structured sparsity patterns should use a SparseMatrixCSC with a correct sparsity pattern for the Hessian. The default is nothing, which means a dense Hessian.

cons_jac_prototype: a prototype matrix matching the type that matches the constraint Jacobian. The default is nothing, which means a dense constraint Jacobian.

cons_hess_prototype: a prototype matrix matching the type that matches the constraint Hessian. This is defined as an array of matrices, where hess[i] is the Hessian w.r.t. the ith output. For example, if the Hessian is sparse, then hess is a Vector{SparseMatrixCSC}. The default is nothing, which means a dense constraint Hessian.

lag_h(res,x,sigma,mu,p) or res=lag_h(x,sigma,mu,p): the Hessian of the Lagrangian, where sigma is a multiplier of the cost function and mu are the Lagrange multipliers multiplying the constraints. This can be provided instead of hess and cons_h to solvers that directly use the Hessian of the Lagrangian.

hess_colorvec: a color vector according to the SparseDiffTools.jl definition for the sparsity pattern of the hess_prototype. This specializes the Hessian construction when using finite differences and automatic differentiation to be computed in an accelerated manner based on the sparsity pattern. Defaults to nothing, which means a color vector will be internally computed on demand when required. The cost of this operation is highly dependent on the sparsity pattern.

cons_jac_colorvec: a color vector according to the SparseDiffTools.jl definition for the sparsity pattern of the cons_jac_prototype.

cons_hess_colorvec: an array of color vector according to the SparseDiffTools.jl definition for the sparsity pattern of the cons_hess_prototype.

When Symbolic Problem Building with ModelingToolkit interface is used the following arguments are also relevant:

syms: the symbol names for the elements of the equation. This should match u0 in size. For example, if u = [0.0,1.0] and syms = [:x, :y], this will apply a canonical naming to the values, allowing sol[:x] in the solution and automatically naming values in plots.

paramsyms: the symbol names for the parameters of the equation. This should match p in size. For example, if p = [0.0, 1.0] and paramsyms = [:a, :b], this will apply a canonical naming to the values, allowing sol[:a] in the solution.

observed: an algebraic combination of optimization variables that is of interest to the user which will be available in the solution. This can be single or multiple expressions.

sys: field that stores the OptimizationSystem.

Defining Optimization Functions via AD

While using the keyword arguments gives the user control over defining all of the possible functions, the simplest way to handle the generation of an OptimizationFunction is by specifying the ADtype which lets the user choose the Automatic Differentiation backend to use for automatically filling in all of the extra functions. For example,

OptimizationFunction(f,AutoForwardDiff())

will use ForwardDiff.jl to define all of the necessary functions. Note that if any functions are defined directly, the auto-AD definition does not overwrite the user's choice.

Each of the AD-based constructors are documented separately via their own dispatches below in the Automatic Differentiation Construction Choice Recommendations section.

iip: In-Place vs Out-Of-Place

For more details on this argument, see the ODEFunction documentation.

specialize: Controlling Compilation and Specialization

For more details on this argument, see the ODEFunction documentation.

Fields

The fields of the OptimizationFunction type directly match the names of the inputs.

Automatic Differentiation Construction Choice Recommendations

The choices for the auto-AD fill-ins with quick descriptions are:

AutoForwardDiff(): The fastest choice for small optimizations

AutoReverseDiff(compile=false): A fast choice for large scalar optimizations

AutoTracker(): Like ReverseDiff but GPU-compatible

AutoZygote(): The fastest choice for non-mutating array-based (BLAS) functions

AutoFiniteDiff(): Finite differencing, not optimal but always applicable

AutoModelingToolkit(): The fastest choice for large scalar optimizations

AutoEnzyme(): Highly performant AD choice for type stable and optimized code

Automatic Differentiation Choice API

The following sections describe the Auto-AD choices in detail.

ADTypes.AutoForwardDiff — Type

AutoForwardDiff{chunksize} <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoForwardDiff(); kwargs...)

This uses the ForwardDiff.jl package. It is the fastest choice for small systems, especially with heavy scalar interactions. It is easy to use and compatible with most Julia functions which have loose type restrictions. However, because it's forward-mode, it scales poorly in comparison to other AD choices. Hessian construction is suboptimal as it uses the forward-over-forward approach.

Compatible with GPUs

Compatible with Hessian-based optimization

Compatible with Hv-based optimization

Compatible with constraints

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not defined via ForwardDiff.

ADTypes.AutoFiniteDiff — Type

AutoFiniteDiff{T1,T2,T3} <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoFiniteDiff(); kwargs...)

This uses FiniteDiff.jl. While not necessarily the most efficient, this is the only choice that doesn't require the f function to be automatically differentiable, which means it applies to any choice. However, because it's using finite differencing, one needs to be careful as this procedure introduces numerical error into the derivative estimates.

Compatible with GPUs

Compatible with Hessian-based optimization

Compatible with Hv-based optimization

Compatible with constraint functions

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not defined via FiniteDiff.

Constructor

AutoFiniteDiff(; fdtype = Val(:forward)fdjtype = fdtype, fdhtype = Val(:hcentral))

fdtype: the method used for defining the gradient

fdjtype: the method used for defining the Jacobian of constraints.

fdhtype: the method used for defining the Hessian

For more information on the derivative type specifiers, see the FiniteDiff.jl documentation.

ADTypes.AutoReverseDiff — Type

AutoReverseDiff <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoReverseDiff(); kwargs...)

This uses the ReverseDiff.jl package. AutoReverseDiff has a default argument, compile, which denotes whether the reverse pass should be compiled. compile should only be set to true if f contains no branches (if statements, while loops) otherwise it can produce incorrect derivatives!

AutoReverseDiff is generally applicable to many pure Julia codes, and with compile=true it is one of the fastest options on code with heavy scalar interactions. Hessian calculations are fast by mixing ForwardDiff with ReverseDiff for forward-over-reverse. However, its performance can falter when compile=false.

Not compatible with GPUs

Compatible with Hessian-based optimization by mixing with ForwardDiff

Compatible with Hv-based optimization by mixing with ForwardDiff

Not compatible with constraint functions

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not defined via ReverseDiff.

Constructor

AutoReverseDiff(; compile = false)

Note: currently, compilation is not defined/used!

ADTypes.AutoZygote — Type

AutoZygote <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoZygote(); kwargs...)

This uses the Zygote.jl package. This is the staple reverse-mode AD that handles a large portion of Julia with good efficiency. Hessian construction is fast via forward-over-reverse mixing ForwardDiff.jl with Zygote.jl

Compatible with GPUs

Compatible with Hessian-based optimization via ForwardDiff

Compatible with Hv-based optimization via ForwardDiff

Not compatible with constraint functions

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not defined via Zygote.

ADTypes.AutoTracker — Type

AutoTracker <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoTracker(); kwargs...)

This uses the Tracker.jl package. Generally slower than ReverseDiff, it is generally applicable to many pure Julia codes.

Compatible with GPUs

Not compatible with Hessian-based optimization

Not compatible with Hv-based optimization

Not compatible with constraint functions

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not defined via Tracker.

ADTypes.AutoModelingToolkit — Type

AutoModelingToolkit <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoModelingToolkit(); kwargs...)

This uses the ModelingToolkit.jl package's modelingtookitize functionality to generate the derivatives and other fields of an OptimizationFunction. This backend creates the symbolic expressions for the objective and its derivatives as well as the constraints and their derivatives. Through structural_simplify, it enforces simplifications that can reduce the number of operations needed to compute the derivatives of the constraints. This automatically generates the expression graphs that some solver interfaces through OptimizationMOI like AmplNLWriter.jl require.

Compatible with GPUs

Compatible with Hessian-based optimization

Compatible with Hv-based optimization

Compatible with constraints

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not generated via ModelingToolkit.

Constructor

AutoModelingToolkit(false, false)

obj_sparse: to indicate whether the objective hessian is sparse.

cons_sparse: to indicate whether the constraints' jacobian and hessian are sparse.

ADTypes.AutoEnzyme — Type

AutoEnzyme <: AbstractADType

An AbstractADType choice for use in OptimizationFunction for automatically generating the unspecified derivative functions. Usage:

OptimizationFunction(f, AutoEnzyme(); kwargs...)

This uses the Enzyme.jl package. Enzyme performs automatic differentiation on the LLVM IR code generated from julia. It is highly-efficient and its ability perform AD on optimized code allows Enzyme to meet or exceed the performance of state-of-the-art AD tools.

Compatible with GPUs

Compatible with Hessian-based optimization

Compatible with Hv-based optimization

Compatible with constraints

Note that only the unspecified derivative functions are defined. For example, if a hess function is supplied to the OptimizationFunction, then the Hessian is not defined via Enzyme.